+++ changelog.md
@@ -154,3 +154,5 @@
 - Changed error message to be more friendly.
 - Added generic error alert.
 - Added error alert in Play and Lead form.
+- Added 'elec_year' in table lead.
+- Added helper function addToSquare.